rename dump.hpp -> cpp-dump.hpp
