+++ test/color_non_variable_test.cpp
@@ -1,4 +1,4 @@
-#include "../dump.hpp"
+#include "../cpp-dump.hpp"
 namespace cp = cpp_dump;
 
 #include <array>
